+++ pkg/providers/iam/data_akamai_iam_account_switch_keys.go
@@ -66,11 +66,11 @@ func (d *accountSwitchKeysDataSource) Schema(_ context.Context, _ datasource.Sch
 		Attributes: map[string]schema.Attribute{
 			"client_id": schema.StringAttribute{
 				Optional:    true,
-				Description: "Unique identifier for each API client. If not provided it assumes your client id.",
+				Description: "A unique identifier for each API client. If not provided, it assumes your client ID.",
 			},
 			"filter": schema.StringAttribute{
 				Optional:    true,
-				Description: "Filters results by accountId or accountName. Enter at least three characters to filter the results by substring.",
+				Description: "Filters results by `account_id` or `account_name`. Enter at least three characters to filter the results by substring.",
 				Validators:  []validator.String{stringvalidator.LengthAtLeast(3)},
 			},
 			"account_switch_keys": schema.ListNestedAttribute{