+++ pkg/providers/iam/provider.go
@@ -100,6 +100,7 @@ func (p *Subprovider) FrameworkDataSources() []func() datasource.DataSource {
 		NewAccessibleGroupsDataSource,
 		NewAccountSwitchKeysDataSource,
 		NewAllowedAPIsDataSource,
+		NewAPIClientDataSource,
 		NewAPIClientsDataSource,
 		NewAuthorizedUsersDataSource,
 		NewBlockedPropertiesDataSource,